% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aco_id_crosswalk.R
\docType{data}
\name{aco_id_crosswalk}
\alias{aco_id_crosswalk}
\title{Crosswalk of ACO IDs to ACO Numbers data from all years.}
\format{
A data frame with 767 rows and 3 variables:
}
\source{
\url{https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SSPACO/index.html}
}
\usage{
aco_id_crosswalk
}
\description{
A dataset providing a crosswalk between ACO ID and ACO NUM. This crosswalk from CMS provides a link between the 5-character ACO ID used in Shared Savings Program public use and research identifiable data files produced in 2019 and later and the 6-character Encrypted ACO ID used in files produced
}
\keyword{datasets}
