% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{cost_metrics}
\alias{cost_metrics}
\title{Helper function to access cost metrics}
\usage{
cost_metrics(df, year = NULL)
}
\arguments{
\item{df}{SSP data}

\item{year}{MSSP performance year.}
}
\value{
Data frame with mssp cost metrics.
}
\description{
Helper function to access cost metrics
}
\examples{
a <- load_puf_file(2022)
cost_metrics(a)
}
