% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_weight_int.R
\name{elnet_weight_int}
\alias{elnet_weight_int}
\title{Elastic Net Weigthing Method}
\usage{
elnet_weight_int(pred_train, dat_train)
}
\arguments{
\item{pred_train}{A list that contain multi-study datasets (can be generated from the simulation function)}

\item{dat_train}{A dataframe that combines all multi-study datasets}
}
\value{
The coeficient for each study using stacking regression with intercept weighting method
}
\description{
Elastic Net Weigthing Method
}
\examples{
lst_ori <- MultiStudySim(k,nk,p,m,mu_x,SIG,pii,mu_beta,sigma_beta)
a <- get_input_data(lst_ori)[[1]]
b <- get_input_data(lst_ori)[[2]]
fit <- lapply(a, function(a) lm(y~.))
b1 <- lapply(fit, function(fit)
                     predict(fit,b[,-which(names(b) \%in\% c('y'))]))
elnet_weight_int(b1,b)
}
\seealso{
[MultiStudySim()]

[get_input_data()]
}
