% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{enhance_puf_file}
\alias{enhance_puf_file}
\title{Applies enhancements to a datafrane containing PUF file data}
\usage{
enhance_puf_file(df, year)
}
\arguments{
\item{df}{Dataframe containing downloaded PUF file.}

\item{year}{MSSP performance year.}
}
\value{
Data frame with mssp data.
}
\description{
Applies enhancements to a datafrane containing PUF file data
}
\examples{
load_puf_file(df, 2016)
}
