% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cms_endpoints.R
\name{get_cms_endpoints}
\alias{get_cms_endpoints}
\title{Generates dataframe with links to CMS data for each year}
\usage{
get_cms_endpoints()
}
\value{
Data frame with URL to get CMs data
}
\description{
Generates dataframe with links to CMS data for each year
}
\examples{
get_cms_endpoints()
}
