% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_data.R
\name{get_input_data}
\alias{get_input_data}
\title{Modification of the simulated results}
\usage{
get_input_data(lst)
}
\arguments{
\item{lst}{The multi-study results can be generated from the simulation function MultiStudySim(k,nk,p,m,mu_x,SIG,pii,mu_beta,sigma_beta)}
}
\value{
Two lists, first list is the multi-study training dataset, second list is the whole training dataset that combines all multi-study datasets
}
\description{
Modification of the simulated results
}
\examples{
lst <- MultiStudySim(k,nk,p,m,mu_x,SIG,pii,mu_beta,sigma_beta)
get_input_data(lst)
}
\seealso{
[MultiStudySim()]
}
