% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_puf_file}
\alias{load_puf_file}
\title{Downloads PUF files from CMS website (https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SSPACO/index.html)}
\usage{
load_puf_file(year, enhance_data = TRUE, standardize_data = FALSE)
}
\arguments{
\item{year}{MSSP performance year.}

\item{enhance_data}{Flag to standardize variables.}
}
\value{
Data frame with mssp data.
}
\description{
Downloads PUF files from CMS website (https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SSPACO/index.html)
}
\examples{
load_puf_file(2016)
}
