% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_functions.R
\name{lookup_aco_by_state}
\alias{lookup_aco_by_state}
\title{Returns list of ACOs in a state}
\usage{
lookup_aco_by_state(df, str)
}
\arguments{
\item{df}{SSP data}

\item{str}{Name of the state}
}
\value{
ACOs where assigned beneficiaries reside in the state
}
\description{
Returns list of ACOs in a state
}
\examples{
a <- load_puf_file(2022)
lookup_aco_by_state(a, "Maine")
}
