% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_class_all.R
\name{method_class_all}
\alias{method_class_all}
\title{Model fitting Method}
\usage{
method_class_all(lst, dat_train, dat_test, learner, nk_train)
}
\arguments{
\item{lst}{A list that contain trianing multi-study datasets (can be generated from the simulation function)}

\item{dat_train}{A dataframe that combines all multi-study datasets}

\item{dat_test}{A list that contains testing multi-study datasets (can be generated from the simulation function)}

\item{learner}{A string that indicates the model fitting method (build under caret):"glmnet" (for Elastic Net)
"gbm" (for gradient boosting), "ranger" (for random forest), "nnet" (for neural network)}

\item{nk_train}{A vector that only needs when using stacking with zero weighting method}
}
\value{
A list that contain the predicted outcome for each multi-study test dataset across differnt weighting methods using selected model fitting method
}
\description{
Model fitting Method
}
\examples{
lst_ori <- MultiStudySim(k,nk,p,m,mu_x,SIG,pii,mu_beta,sigma_beta)
a <- get_input_data(lst_ori)[[1]]
b <- get_input_data(lst_ori)[[2]]
test <- lapply(MultiStudySim(k,nk,p,m,mu_x,SIG,pii,mu_beta,sigma_beta),'[[',1)
l <- 'glmnet'
nk_train <- nk
method_class_all(a,b,test,l)
method_class_all(a,b,test,l,nk)

}
\seealso{
[MultiStudySim()]

[get_input_data()]
}
