% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssp_all_years.R
\docType{data}
\name{mssp_all_years}
\alias{mssp_all_years}
\title{MSSP data from all years.}
\format{A data frame with 1849 rows and 166 variables:}
\source{
\url{https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SSPACO/index.html}
}
\usage{
mssp_all_years
}
\description{
A dataset containing information on ACO performance across all years. Values have been standardized using the variables published in the 2017 SSP PUF file.
}
\keyword{datasets}
