% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationCode.R
\name{MultiStudySim}
\alias{MultiStudySim}
\title{Simulation function for Multi-Study Dataset.}
\usage{
MultiStudySim(k, nk, p, m, mu_x, SIG, pii, mu_beta, sigma_beta)
}
\arguments{
\item{k}{A number indicating the number of Studies.}

\item{nk}{A vector indicating the number of observations within each study}

\item{p}{A number indicating the number of covariates within each study. Assuming equal covariates for each study right now}

\item{m}{A number indicating the number of mix component}

\item{mu_x}{A k by p matrix indicaitng the mean of i_th covairates for i=1,...,p, and in j_th study for j=1,...,k}

\item{SIG}{A p by p matrix indicating the variance-covariance matrix of mu_x}

\item{pii}{A m by 1 vector indicating the mixed probability of choosing a specific cluster of beta}

\item{mu_beta}{A m by p matrix indicating the mean of coefficient of i_th covariates for i=1,...,p, and q_th cluster for q=1,...,m}

\item{sigma_beta}{A m by p matrix indicating the variance of coefficient of i_th covariates for i=1,...,p, and q_th cluster for q=1,...,m}
}
\value{
Three lists of multi-study datasets. Three sublists within each \code{k} studies.
First: Simulated Data
Second: beta Coefficients
Third: integer indicating the selectd component
}
\description{
Simulation function for Multi-Study Dataset.
}
\examples{
k <- 10
nk <- rep(100,k)
p <- 3
m <- 4
mu_x <- matrix(runif(k*p,1,20),nrow = k,ncol = p,byrow = T)
SIG <- MCMCpack::riwish(v = p+1, S = diag(1, p,p))
pii <- matrix(runif(m,0,1),nrow = m,ncol = 1,byrow = T)
mu_beta <- matrix(runif(m*p,1,20),nrow = m,ncol = p,byrow = T)
sigma_beta <- matrix(runif(m*p,1,1.1),nrow = m,ncol = p,byrow = T)
MultiStudySim(k,nk,p,m,mu_x,SIG,pii,mu_beta,sigma_beta)
}
\author{
Xutao Wang, \email{xutaow@bu.edu}
}
