% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_expenditures}
\alias{profile_expenditures}
\title{Profiles utilization metrics vs. national sample}
\usage{
profile_expenditures(df, aco_num, year = NULL)
}
\arguments{
\item{df}{SSP data}

\item{aco_num}{ACO Number.}

\item{year}{MSSP performance year.}
}
\value{
Data frame with mssp data.
}
\description{
Profiles utilization metrics vs. national sample
}
\examples{
profile_utilization(df, 2016)
}
