% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_quality}
\alias{profile_quality}
\title{Profiles quality measures vs. national sample}
\usage{
profile_quality(df, aco_id, year = NULL)
}
\arguments{
\item{df}{SSP data}

\item{aco_id}{ACO ID}

\item{year}{MSSP performance year.}
}
\value{
Data frame with mssp data.
}
\description{
Profiles quality measures vs. national sample
}
\examples{
profile_quality(df, "A95164", 2016)
}
