% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_avg.R
\name{simp_avg}
\alias{simp_avg}
\title{Simple Average Weigthing Method}
\usage{
simp_avg(pred_test)
}
\arguments{
\item{pred_test}{A list that cotain the predicted results using the model trained from each single study}
}
\value{
A vector of predicted outcome that evaluated taking the mean value of predicted outcome from \code{k} studies
}
\description{
Simple Average Weigthing Method
}
\examples{
pred_test <- MultiStudySim(k,nk,p,m,mu_x,SIG,pii,mu_beta,sigma_beta)
}
