% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_weight_combine.R
\name{stack_weight_combine}
\alias{stack_weight_combine}
\title{General Stacking Regression Weigthing Method}
\usage{
stack_weight_combine(pred_train, dat_train)
}
\arguments{
\item{pred_train}{A list that contain multi-study datasets (can be generated from the simulation function)}

\item{dat_train}{A dataframe that combines all multi-study datasets}
}
\value{
The coeficient for each study using general stacking regression weighting method
}
\description{
General Stacking Regression Weigthing Method
}
\examples{
lst_ori <- MultiStudySim(k,nk,p,m,mu_x,SIG,pii,mu_beta,sigma_beta)
a <- get_input_data(lst_ori)[[1]]
b <- get_input_data(lst_ori)[[2]]
fit <- lapply(a, function(a) lm(y~.))
b1 <- lapply(fit, function(fit)
                     predict(fit,b[,-which(names(b) \%in\% c('y'))]))
stack_weight_combine(b1,b)
}
\seealso{
[MultiStudySim()]

[get_input_data()]
}
