% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{trend_expenditures}
\alias{trend_expenditures}
\title{Trends expenditures over time}
\usage{
trend_expenditures(df, aco_num)
}
\arguments{
\item{df}{SSP data}

\item{aco_num}{ACO Number.}
}
\value{
Data frame with mssp data.
}
\description{
Trends expenditures over time
}
\examples{
trend_savings(df, "A95164")
}
