% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mssp_all_years.R
\docType{data}
\name{aco_id_crosswalk}
\alias{aco_id_crosswalk}
\title{Crosswalk of ACO IDs to ACO Numbers data from all years.}
\format{
A data frame with 767 rows and 3 variables:
}
\source{
\url{https://www.cms.gov/Research-Statistics-Data-and-Systems/Downloadable-Public-Use-Files/SSPACO/index.html}
}
\usage{
aco_id_crosswalk
}
\description{
A dataset providing a crosswalk between ACO ID and ACO NUM.
}
\keyword{datasets}
