% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_multi_year_db}
\alias{load_multi_year_db}
\title{Downloads PUF files from CMS website from multiple years and integrates into a single dataset.}
\usage{
load_multi_year_db(verbose = FALSE)
}
\arguments{
\item{verbose}{Provides detail on download process if set to TRUE}
}
\value{
Data frame with mssp data from all years.
}
\description{
Downloads PUF files from CMS website from multiple years and integrates into a single dataset.
}
\examples{
df <- load_multi_year_db()
}
