% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_puf_file}
\alias{load_puf_file}
\title{Downloads PUF files from CMS.}
\usage{
load_puf_file(year)
}
\arguments{
\item{year}{MSSP performance year.}
}
\value{
Data frame with mssp data.
}
\description{
Downloads PUF files from CMS.
}
\examples{
load_puf_file(2016)
}
