% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{lookup_aco_by_state}
\alias{lookup_aco_by_state}
\title{Looks up the ACO Num with an ACO name}
\usage{
lookup_aco_by_state(df, str)
}
\arguments{
\item{df}{SSP data}

\item{str}{Name of the state}
}
\value{
ACOs where assigned beneficiaries reside in the state
}
\description{
Looks up the ACO Num with an ACO name
}
\examples{
lookup_aco_num("Massachusetts")
}
