% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{lookup_aco_num}
\alias{lookup_aco_num}
\title{Looks up the ACO Num with an ACO name}
\usage{
lookup_aco_num(str)
}
\arguments{
\item{str}{Search string with the name of the ACO}
}
\value{
Potential matches
}
\description{
Looks up the ACO Num with an ACO name
}
\examples{
lookup_aco_num("Mercy")
}
