% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_savings}
\alias{profile_savings}
\title{Profiles contract measures vs. national sample}
\usage{
profile_savings(df, aco_num, year = NULL)
}
\arguments{
\item{df}{SSP data}

\item{aco_num}{ACO Number.}

\item{year}{MSSP performance year.}
}
\value{
Data frame with mssp data.
}
\description{
Profiles contract measures vs. national sample
}
\examples{
profile_savings(df, "A95164", 2016)
}
