% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_savings}
\alias{profile_savings}
\title{Profiles contract measures vs. national sample}
\usage{
profile_savings(df, aco_id, year = NULL)
}
\arguments{
\item{df}{SSP data}

\item{aco_id}{ACO ID}

\item{year}{MSSP performance year.}
}
\value{
Data frame with mssp data.
}
\description{
Profiles contract measures vs. national sample
}
\examples{
a <- load_enhanced_puf_file(2022)
profile_savings(a, "A1490")
}
