% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{profile_utilization}
\alias{profile_utilization}
\title{Profiles utilization metrics vs. national sample}
\usage{
profile_utilization(df, aco_id, year = NULL)
}
\arguments{
\item{df}{SSP data}

\item{year}{MSSP performance year.}

\item{aco_num}{ACO Number.}
}
\value{
Data frame with mssp data.
}
\description{
Profiles utilization metrics vs. national sample
}
\examples{
profile_utilization(df, "A95164", 2016)
}
