% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{quality_metrics}
\alias{quality_metrics}
\title{Helper function to access quality metrics}
\usage{
quality_metrics(df, year = NULL)
}
\arguments{
\item{df}{SSP data}

\item{year}{MSSP performance year.}
}
\value{
Data frame with mssp quality metrics.
}
\description{
Helper function to access quality metrics
}
\examples{
a <- load_enhanced_puf_file(2022)
quality_metrics(a, 2022)
}
