% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{savings_metrics}
\alias{savings_metrics}
\title{Helper function to access savings metrics}
\usage{
savings_metrics(df, year = NULL)
}
\arguments{
\item{df}{SSP data}

\item{year}{MSSP performance year.}
}
\value{
Data frame with mssp savings metrics.
}
\description{
Helper function to access savings metrics
}
\examples{
savings_metrics(df, 2016)
}
