% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{trend_expenditures}
\alias{trend_expenditures}
\title{Trends expenditures over time}
\usage{
trend_expenditures(aco_id, df = NULL)
}
\arguments{
\item{df}{Optional dataframe with MSSP data. If not supplied, uses mssp_all_years package in dataset}

\item{aco_num}{ACO Number.}
}
\value{
Data frame with expenditure metrics over time.
}
\description{
Trends expenditures over time
}
\examples{
trend_expenditures("A95164")
}
