% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{trend_savings}
\alias{trend_savings}
\title{Trends contract savings over time}
\usage{
trend_savings(aco_num, df = NULL)
}
\arguments{
\item{aco_num}{ACO Number.}

\item{df}{Optional dataframe with MSSP data. If not supplied, uses mssp_all_years package in dataset}
}
\value{
Data frame with mssp data.
}
\description{
Trends contract savings over time
}
\examples{
trend_savings("A95164")
}
