% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{trend_utilization}
\alias{trend_utilization}
\title{Trends utilization over time}
\usage{
trend_utilization(aco_id, df = NULL)
}
\arguments{
\item{df}{Optional dataframe with MSSP data. If not supplied, uses mssp_all_years package in dataset}

\item{aco_num}{ACO Number.}
}
\value{
Data frame with utilization metrics over time.
}
\description{
Trends utilization over time
}
\examples{
trend_savings("A95164")
}
