% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{utilization_metrics}
\alias{utilization_metrics}
\title{Helper function to access utilization metrics}
\usage{
utilization_metrics(df, year = NULL)
}
\arguments{
\item{df}{SSP data}

\item{year}{MSSP performance year.}
}
\value{
Data frame with mssp utilization metrics.
}
\description{
Helper function to access utilization metrics
}
\examples{
a <- load_enhanced_puf_file(2022)
utilization_metrics(a, 2016)
}
